\name{Cancerrate}
\alias{Cancerrate}
\docType{data}
\title{Breast Cancer Data}
\description{
Age-specific breast cancer rates for Australian females with 9 age groups (45-49, 50-54, 55-59, 60-64, 65-69, 70-74, 75-79, 80-84, 85+) from 1921 to 2001. 
}
\usage{
data(Cancerrate)
}
\format{
An object of class \code{fts}.
}
\examples{
plot(Cancerrate)
}
\source{
Australian Institute of Health and Welfare (AIHW) website at \url{http://www.aihw.gov.au/cancer/data/index.cfm}.
}
\references{
B. Erbas and R. J. Hyndman and D. Gertig (2007) "Forecasting age-specific breast cancer mortality using functional data models", \emph{Statistics in Medicine}, \bold{26}(2), 458-470.

R. J. Hyndman and M. S. Ullah (2007) "Robust forecasting of mortality and fertility rates: A functional data approach", \emph{Computational Statistics and Data Analysis}, \bold{51}(10), 4942-4956.
}
\keyword{datasets}

