\name{ECBYieldcurve}
\alias{ECBYieldcurve}
\docType{data}
\title{
Yield curve data spot rate 
}
\description{
Provided by European Central Bank, this data set contains daily yield curve spot rate from 29/12/2006 to 24/07/2009 for government bond, nominal, all triple AAA issued companies,with maturity term at 3, 6 months and 1 to 30 years.
}
\usage{
data(ECBYieldcurve)
}
\format{
An object of class \code{fts}.
}
\note{
We thank Mr Sergio S. Guirreri for the permission to re-distribute this data set.
}
\source{
European Central Bank: \url{http://www.ecb.europa.eu/stats/money/yc/html/index.en.html}.
}
\examples{
plot(ECBYieldcurve)
}
\keyword{datasets}

