\name{fds-package}
\alias{fds-package}                    
\docType{package}
\title{
Functional data sets
}
\description{
This package contains a list of functional time series, sliced functional time series, and functional data sets. 
Functional time series is a special type of functional data observed over time. 
Sliced functional time series is a special type of functional time series with a time variable observed over time. 
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <hanlin.shang@anu.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2010) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Journal of Computational and Graphical Statistics}, \bold{19}(1), 29-45.

R. J. Hyndman and H. L. Shang (2009) "Forecasting functional time series (with discussion)", \emph{Journal of the Korean Statistical Society}, \bold{38}(3), 199-221.

H. L. Shang and R. J. Hyndman (2009) "Nonparametric time series forecasting with dynamic updating", \emph{Mathematics and Computers in Simulation}, \bold{81}(7), 1310-1324.
}
\keyword{package}


