\name{FedYieldcurve}
\alias{FedYieldcurve}
\docType{data}
\title{
Federal Reserve interest rate
}
\description{
This data set contains monthly interest rate of the Federal Reserve from January 1982 to June 2009.
}
\usage{
data(FedYieldcurve)
}
\format{
An object of class \code{fts}.
}
\source{
Board of Governors of the Federal Reserve System at \url{https://www.federalreserve.gov/}.
}
\note{
We thank Mr Sergio S. Guirreri for the permission to re-distribute this data set.
}
\examples{
plot(FedYieldcurve)
}
\keyword{datasets}

