\name{SOI}
\alias{SOI}
\docType{data}
\title{
Annual Southern Oscillation Index (SOI) for the period 1900-2004.
}
\description{
Annual measures on Southern Oscillation Index (SOI): observed annual cycles in period 1900-2004.
}
\usage{
data(SOI)
}
\format{
An object of class \code{fts}.
}
\source{
The data are available at the Australian Meteorological Office (\url{http://www.environment.gov.au}).
}
\examples{
plot(SOI)
}
\keyword{datasets}
