\name{fds-package}
\alias{fds-package}                    
\docType{package}
\title{
Functional data sets
}
\description{
This package contains a list of functional time series, sliced functional time series, and functional data sets. 
Functional time series is a special type of functional data observed over time. 
Sliced functional time series is a special type of functional time series with a time variable observed over time. 
}
\author{
Han Lin Shang and Rob J Hyndman

Maintainer: Han Lin Shang <Hanlin.Shang@buseco.monash.edu.au>
}
\references{
R. J. Hyndman and H. L. Shang. (2009) "Rainbow plots, bagplots, and boxplots for functional data", \emph{Journal of Computational and Graphical Statistics}, \bold{in press}.

R. J. Hyndman and H. L. Shang (2009) "Forecasting functional time series (with discussion)", \emph{Journal of the Korean Statistical Society}, \bold{38}(3), 199-221.

H. L. Shang and R. J. Hyndman (2009) "Nonparametric time series forecasting with dynamic updating", In R. S. Anderssen, R. D. Braddock and L.T.H. Newham (eds), 18th World IMACS Congress and MODSIM09 International Congress on Modelling and Simulation. Modelling and Simulation Society of Australia and New Zealand and International Association for Mathematics and Computers in Simulation, July 2009, pp. 1552-1558. ISBN: 978-0-9758400-7-8. http://www.mssanz.org.au/modsim09/D11/shang.pdf
}
\keyword{package}


